Experiments/73_Mcp/django-mcp-ecommerce/mcp/views.py
from django.shortcuts import render
from rest_framework import viewsets, status
from rest_framework.decorators import action
from rest_framework.response import Response
from .models import Product
from .serializers import ProductSerializer

# Create your views here.

class ProductViewSet(viewsets.ModelViewSet):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer

    @action(detail=False, methods=['get'])
    def get_products(self, request):
        products = self.get_queryset()
        serializer = self.get_serializer(products, many=True)
        return Response(serializer.data)

    @action(detail=True, methods=['get'])
    def recommend_product(self, request, pk=None):
        product = self.get_object()
        # Here you can implement your recommendation logic
        # For now, we'll just return the product details
        serializer = self.get_serializer(product)
        return Response(serializer.data)
