Experiments/73_Mcp/django-mcp-ecommerce/requirements.txt
django 
djangorestframework 
django-cors-headers